+++ src/main/java/net/sf/jabref/logic/groups/AbstractGroup.java
@@ -61,18 +61,18 @@ public abstract class AbstractGroup implements SearchMatcher {
      * @throws Exception If an error occured and a group could not be created, e.g.
      *                   due to a malformed regular expression.
      */
-    public static AbstractGroup fromString(String s, int version) throws Exception {
+    public static AbstractGroup fromString(String s) throws Exception {
         if (s.startsWith(KeywordGroup.ID)) {
-            return KeywordGroup.fromString(s, version);
+            return KeywordGroup.fromString(s);
         }
         if (s.startsWith(AllEntriesGroup.ID)) {
-            return AllEntriesGroup.fromString(s, version);
+            return AllEntriesGroup.fromString(s);
         }
         if (s.startsWith(SearchGroup.ID)) {
-            return SearchGroup.fromString(s, version);
+            return SearchGroup.fromString(s);
         }
         if (s.startsWith(ExplicitGroup.ID)) {
-            return ExplicitGroup.fromString(s, version);
+            return ExplicitGroup.fromString(s);
         }
         return null; // unknown group
     }